refactor: 리사이클러뷰용 데이터를 각 presertation 패키지로 이동
