+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/adapter/FriendAdapter.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import org.sopt.dosopttemplate.data.Friend
+import org.sopt.dosopttemplate.presentation.home.data.Friend
 import org.sopt.dosopttemplate.databinding.ItemFriendListBinding
 import org.sopt.dosopttemplate.presentation.home.viewholder.FriendViewHolder
