+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/viewholder/FriendViewHolder.kt
@@ -1,7 +1,7 @@
 package org.sopt.dosopttemplate.presentation.home.viewholder
 
 import androidx.recyclerview.widget.RecyclerView
-import org.sopt.dosopttemplate.data.Friend
+import org.sopt.dosopttemplate.presentation.home.data.Friend
 import org.sopt.dosopttemplate.databinding.ItemFriendListBinding
 
 class FriendViewHolder(private val binding: ItemFriendListBinding) :
