+++ app/src/main/java/org/sopt/dosopttemplate/presentation/home/viewholder/MyProfileViewHolder.kt
@@ -1,7 +1,7 @@
 package org.sopt.dosopttemplate.presentation.home.viewholder
 
 import androidx.recyclerview.widget.RecyclerView
-import org.sopt.dosopttemplate.data.MyProfile
+import org.sopt.dosopttemplate.presentation.home.data.MyProfile
 import org.sopt.dosopttemplate.databinding.ItemMyProfileBinding
 
 class MyProfileViewHolder(private val binding: ItemMyProfileBinding) :
